backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
dataclasses-json==0.6.7
dnspython==2.4.2
fastapi==0.104.1
google-ai-generativelanguage==0.6.6
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
httplib2==0.31.0
httptools==0.6.4
idna==3.10
IMAPClient==2.3.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
mail-parser==3.15.0
marshmallow==3.26.1
more-itertools==10.8.0
mypy_extensions==1.1.0
nylas==6.5.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pynput==1.8.1
pyparsing==3.2.5
pystray==0.19.5
pywin32==311
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchdog==4.0.1
watchfiles==1.1.0
websockets==15.0.1
